+++ src/desktop/core/syncAppSettingsToKintone.ts
